requirements.txt
python-dotenv
pandas
openai
anthropic
moviepy
pydub